refactor: auth utils directory to include club and event authentication utils
