+++ lambda/api/clubs/clubId/events/post/post.go
@@ -3,7 +3,7 @@ package main
 import (
 	"cdk-infrastructure/database/models"
 	gateway_helpers "cdk-infrastructure/gateway/helpers"
-	authentication_utils "cdk-infrastructure/utils/authentication"
+	authentication_utils "cdk-infrastructure/utils/auth"
 	validation_error "cdk-infrastructure/utils/errors"
 	"cdk-infrastructure/utils/query_client"
 	"context"
