+++ lambda/api/me/clubs/get.go
@@ -4,7 +4,7 @@ import (
 	// models "cdk-infrastructure/database/models"
 	"cdk-infrastructure/database/models"
 	gateway_helpers "cdk-infrastructure/gateway/helpers"
-	authentication_utils "cdk-infrastructure/utils/authentication"
+	authentication_utils "cdk-infrastructure/utils/auth"
 	"cdk-infrastructure/utils/query_client"
 	"context"
 	"errors"
